deleting files now closes tabs

+++ src/main/VR_File_Tree.rb
@@ -124,19 +124,26 @@ class VR_File_Tree < VR::FileTreeView
     return unless evt.keyval == 65535 #delete
     return unless file_name = get_selected_path()
     return unless alert("Delete:   <b>" + File.basename(file_name) + "</b> ?" , :button_yes => "Delete", :button_no=>"Cancel", :parent=>self, :headline => "Delete FIle?")  
-    if File.exist?(file_name)
-      @main.tabs.destroy_file_tab(file_name)
+
+#    if File.directory?(file_name) and file_name != Dir.pwd #root!
+#      delete_me = @main.docs.select { |doc| file_name ~= doc.full_path_file }
+
+#      FileUtils.remove_dir(file_name, true)
+
+    if File.exist?(file_name) and file_name != Dir.pwd #root!
+      delete_me =  @main.tabs.docs.select { |d| d.full_path_file.start_with?(file_name) }
+      delete_me.each { |d| @main.tabs.destroy_file_tab(d.full_path_file) }
+      delete_selected()
+#      @main.tabs.destroy_file_tab(file_name)
       #this is where the nasty error occurs.  If you save_as a file, then delete the old
       # one, the following line crashes the prgogram.
-      @main.file_tree.model.clear
+#      @main.file_tree.model.clear
       FileUtils.rm_rf(file_name)
 
 #      File.delete(file_name)
-    elsif File.directory?(file_name) and file_name != Dir.pwd #root!
-      FileUtils.remove_dir(file_name, true)
-    end
-    @main.file_tree.refresh()
 #    delete_selected()
+    end
+
   end
